[Feat] SessionDetailEntity 파일로 분리
